refactor(engine): peel more merger + heartbeat recovery satellites
Wave 3 of package organization:
- merger-errors: verification/abort error classes
- merger-owned-landed: ownership classification + Fusion-Task-Id trailer
- merger-conflict-resolution: conflict classify/auto-resolve helpers
- agent-heartbeat-error-recovery: durable error-recovery budget helpers

Preserve public re-exports from merger.ts / agent-heartbeat.ts. Fix
isTrivialWhitespaceConflict empty-stdout catch parity with monofile.
